requirements.txt
streamlit
pandas
requests
matplotlib
seaborn
google-cloud-bigquery
pymongo
streamlit-aggrid
watchdog